Ensure normalized remotes have .git
